(svn r15751) Fix: compilation (svn r15751) Backport: improvement of string helper functions

+++ src/shared/debug.cpp
@@ -137,7 +137,7 @@ void SetDebugString(const char *s)
 		if (p != NULL) {
 			*p = v;
 		} else {
-			printf("Unknown debug level '%.*s'", s - t, t);
+			printf("Unknown debug level '%.*s'", (int)(s - t), t);
 			return;
 		}
 	}
